chore(jest): remove `jest.config.ts` file (#234)
The `jest.config.ts` file was removed as it is no longer needed for the project. This file contained the configuration settings for Jest, the testing framework used in the project. Since the file is no longer necessary, it was deleted to simplify the project structure and reduce unnecessary files.